Move FCN8sDepthPrediction chainer models to jsk_recognition_utils
